Create raylib_example_code.txt
this is some code that I have so far for raylib, not sure how to get all of it set up with github at the moment, and variable names are bound to change, so I'm just leving it here for the moment. 

Here are instructions i used for setting up raylib on my Mac: https://github.com/raysan5/raylib/wiki/Compile-for-OSX 
then you can just copy/paste this code into a new project to try it out.