Update raylib_example_code
this can take input and display it on the board, but it doesn't check for duplicate numbers. a checkImpossibility function would likely be nested in here and would display an error onscreen the moment an impossible board was entered. since it stores the numbers to display on the board in an array it could then pass that array to the part of the function that processes the board, and then that could output a file.