fixed some issues that crop up with longer problems
there were some big mistakes that somehow didn't cause a problem with boards that could be solved in 4 moves. 

it gets into an infinite loop of D,R,L,U,D,R,L,... which I think is due to the fact that it still adds boards to the queue that it has already explored. 

https://www.cs.princeton.edu/courses/archive/spr10/cos226/assignments/8puzzle.html mentions this: 
"After implementing best-first search, you will notice one annoying feature: states corresponding to the same board position are enqueued on the priority queue many times. To prevent unnecessary exploration of useless states, when considering the neighbors of a state, don't enqueue the neighbor if its board position is the same as the previous state."