No longer infinite loop, but makes illegal moves
So I fixed the infinite loop problem, but it makes some illegal moves. I think the illegal moves are a result of the method I used to keep track of moves; I had a hunch that would be the case when I wrote it.

also, I forgot the scoring function is the Manhattan distance + the number of previous moves, which is what initially fixed the infinite loop.